+++ Sources/Error.swift
@@ -1,0 +1,5 @@
+public enum Error: Swift.Error {
+    case invalidRequest(Swift.Error)
+    case invalidResponse(Swift.Error)
+    case notConnected(Swift.Error)
+}