+++ Sources/HTTPMethod.swift
@@ -1,0 +1,10 @@
+enum HTTPMethod: String {
+    case options = "OPTIONS"
+    case get = "GET"
+    case head = "HEAD"
+    case post = "POST"
+    case put = "PUT"
+    case delete = "DELETE"
+    case trace = "TRACE"
+    case connect = "CONNECT"
+}